+++ packages/core/src/main/mls/MLSService/MLSService.ts
@@ -28,7 +28,7 @@ import {
 import {APIClient} from '@wireapp/api-client';
 import {QualifiedUsers} from '../../conversation';
 import {Decoder, Encoder} from 'bazinga64';
-import {MLSConfig} from '../../Account';
+import type {MLSConfig} from '../types';
 
 //@todo: this function is temporary, we wait for the update from core-crypto side
 //they are returning regular array instead of Uint8Array for commit and welcome messages
